backend/requirements/base.txt
# Base dependencies
Django>=5.1.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=23.5
python-dotenv>=1.0.0
typing_extensions>=4.12.2
psycopg2-binary>=2.9.10  # PostgreSQL database adapter
PyJWT>=2.10.1
pytz>=2024.2